moved rules.js from components/tabscreens to screens folder
